# Allow one timing to cancel others